pyproject.toml
[project]
name = "nikki"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "langchain >=0.3.7",
    "streamlit >=1.35.0",
    "lancedb >=0.16.0",
    "chromadb >=0.5.20",
    "langchain-community >=0.3.0",
    "python-dotenv >=1.0.1",
    "sentence-transformers >=3.0.1",
    "streamlit-chat >=0.1.1",
    "langchain-huggingface>=0.0.3",
    "markdown >=3.6",
    "weasyprint >=62.3",
    "jinja2 >=3.1.4",
    "pango >=0.0.1",
    "pydyf >=0.11.0",
    "cffi >=1.16.0",
    "html5lib >=1.1",
    "tinycss2 >=1.3.0",
    "cssselect2 >=0.7.0",
    "pyphen >=0.15.0",
    "pillow >=10.4.0",
    "fonttools >=4.53.1",
    "gobject >=0.1.0",
    "markdown2 >=2.5.0",
    "tantivy >=0.22.0",
    "llama-cpp-python >=0.2.88",
    "safetensors >=0.4.4",
    "langchain-cli>=0.0.33",
    "langchain-chroma>=0.1.4",
]
